crewai_agents/requirements.txt
crewai==0.41.1
openai>=1.12.0
wandb>=0.16.0
weave>=0.50.0
python-dotenv>=1.0.0
pydantic>=2.5.0
spacy>=3.7.0
requests>=2.31.0
google-generativeai>=0.3.0
